refactor: complete migration to Shardate class and update tests
- Rename Reader class to Shardate
- Move implementation to shardate.py
- Update __init__.py imports
- Remove old read.py file
- Update test imports and references
