+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "shardate"
-version = "2025.9.8"
+version = "2025.9.9"
 description = "A lightweight Python library for efficiently reading year-month-day partitioned Parquet datasets."
 readme = "README.md"
 requires-python = ">=3.12"
