+++ src/shardate/__init__.py
@@ -1,1 +1,3 @@
 """Shardate: Fast reader for YMD-partitioned Parquet files."""
+from .shardate import Shardate  # noqa: F401
+all = ["Shardate"]
