+++ tests/test_read.py
@@ -4,7 +4,7 @@ from typing import Iterable
 import pytest
 from pyspark.sql import SparkSession
 
-from shardate.read import Reader
+from shardate.shardate import Shardate
 from shardate.dates import all_dates_between
 from shardate.utils import date_col
 
@@ -56,26 +56,26 @@ def path(spark, start_date, end_date, tmp_path_factory):
 
 
 def test_read_by_date(path: str, target_date: date):
-    df = Reader(path).read_by_date(target_date)
+    df = Shardate(path).read_by_date(target_date)
     dates = [row["date"] for row in sorted(df.select(date_col()).distinct().collect())]
     assert len(dates) == 1
     assert dates[0] == target_date
 
 
 def test_read_between(path: str, start_date: date, end_date: date):
-    df = Reader(path).read_between(start_date, end_date)
+    df = Shardate(path).read_between(start_date, end_date)
     dates = [row["date"] for row in sorted(df.select(date_col()).distinct().collect())]
     assert dates == list(all_dates_between(start_date, end_date))
 
 
 def test_read_by_dates(path: str, target_dates: Iterable[date]):
-    df = Reader(path).read_by_dates(target_dates)
+    df = Shardate(path).read_by_dates(target_dates)
     dates = [row["date"] for row in sorted(df.select(date_col()).distinct().collect())]
     assert dates == sorted(target_dates)
 
 
 def test_read_eoms_between(path: str, start_date: date, end_date: date):
-    df = Reader(path).read_eoms_between(start_date, end_date)
+    df = Shardate(path).read_eoms_between(start_date, end_date)
     dates = [row["date"] for row in sorted(df.select(date_col()).distinct().collect())]
     expected_dates = [
         dt for dt in all_dates_between(start_date, end_date) if dt.day == 31
